add validation in delete and purchaise product.

+++ bookstore/src/app/product/detail/detail.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, ViewChild, ElementRef } from "@angular/core";
 import { ProductService } from "../product.service";
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { IProduct } from "src/app/shared/interfaces/product";
 import { IProducts } from "src/app/shared/interfaces/products";
 import { AuthService } from "src/app/authentication/auth.service";
@@ -23,7 +23,8 @@ export class DetailComponent implements OnInit {
   constructor(
     private productService: ProductService,
     private authService: AuthService,
-    private activatedRoute: ActivatedRoute
+    private activatedRoute: ActivatedRoute,
+    private router: Router
   ) {}
 
   ngOnInit() {
@@ -34,30 +35,46 @@ export class DetailComponent implements OnInit {
       });
   }
 
-  isHasCredentials() {    
-    let h = this.authService.checkIfLogged() && this.authService.roles;    
+  isHasCredentials() {
+    let h = this.authService.checkIfLogged() && this.authService.roles;
     if (h) {
       return true;
     }
     return false;
   }
-  makePurchase() {
-    debugger
-    this.productService
-      .purchase(+this.amountInput.nativeElement.value)
-      .subscribe(() => {
-        debugger
-        console.log('its ok!')
-        this.productService.loading();
-        this.amountInput.nativeElement.value = "";
-      });
+  isAvailable() {
+    let test = this.selectedProduct.quantity > 0;
+    if (test) {
+      return true;
+    }
+    return false;
   }
-  removeProduct() {  
-    if (this.isHasCredentials()) {
+  makePurchase() {    
+    if (this.authService.authtoken) {
+      let result = confirm("Do you want this product?");
+      debugger
+      if (result) {
+        this.productService
+          .purchase(+this.amountInput.nativeElement.value)
+          .subscribe(() => {
+            debugger;
+            console.log("its ok!");
+            this.productService.loading();
+            this.amountInput.nativeElement.value = "";
+          });
+      }else {
+        this.router.navigateByUrl('');
+      }
+    } else {
+      this.router.navigateByUrl("/login");
+    }
+  }
+  removeProduct() {
+    let test = confirm("Are are you shure, this will delete product from DB?");
+    if (this.isHasCredentials() && test) {
       this.productService.remove().subscribe(() => {
         this.productService.loading();
       });
     }
- 
   }
 }
